requirements.txt
# Stable Diffusion Framework for KITTI-360 Satellite-to-Frontview Generation

# Core Stable Diffusion
diffusers==0.27.2
transformers==4.38.2
accelerate==0.27.2
bitsandbytes==0.43.0

# PyTorch
torch==2.2.0
torchvision==0.17.0

# Project dependencies
numpy==1.26.4
opencv-python==4.9.0.80
Pillow==10.3.0
scipy==1.11.4
pyyaml==6.0.1
tqdm==4.66.2

# Logging and visualization
wandb==0.16.4
tensorboardX==2.6.2.2

# Optional
einops==0.7.0
xformers==0.0.24.post1
